requirements.txt
celery==5.0.5
celery-singleton==0.3.1
croniter==1.0.13
Django==3.2.3
django-celery==3.1.17
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-extensions==3.1.3
django-pandas==0.6.4
django-ratelimit==4.0.0
django-rest-framework==0.1.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
dnspython==2.1.0
gunicorn==20.1.0
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
matplotlib-inline==0.1.2
notebook==6.4.0
numpy==1.20.3
pandas==1.2.4
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pymongo==3.11.4
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
